🛠️ first round of apiRouter refactor
