+++ adminSiteServer/apiRoutes/tagGraph.ts
@@ -1,0 +1,60 @@
+import { JsonError, FlatTagGraph } from "@ourworldindata/types"
+import { checkIsPlainObjectWithGuard } from "@ourworldindata/utils"
+import { apiRouter } from "../apiRouter.js"
+import {
+    getRouteWithROTransaction,
+    postRouteWithRWTransaction,
+} from "../functionalRouterHelpers.js"
+import * as db from "../../db/db.js"
+import * as lodash from "lodash"
+
+getRouteWithROTransaction(
+    apiRouter,
+    "/flatTagGraph.json",
+    async (req, res, trx) => {
+        const flatTagGraph = await db.getFlatTagGraph(trx)
+        return flatTagGraph
+    }
+)
+
+postRouteWithRWTransaction(apiRouter, "/tagGraph", async (req, res, trx) => {
+    const tagGraph = req.body?.tagGraph as unknown
+    if (!tagGraph) {
+        throw new JsonError("No tagGraph provided", 400)
+    }
+
+    function validateFlatTagGraph(
+        tagGraph: Record<any, any>
+    ): tagGraph is FlatTagGraph {
+        if (lodash.isObject(tagGraph)) {
+            for (const [key, value] of Object.entries(tagGraph)) {
+                if (!lodash.isString(key) && isNaN(Number(key))) {
+                    return false
+                }
+                if (!lodash.isArray(value)) {
+                    return false
+                }
+                for (const tag of value) {
+                    if (
+                        !(
+                            checkIsPlainObjectWithGuard(tag) &&
+                            lodash.isNumber(tag.weight) &&
+                            lodash.isNumber(tag.parentId) &&
+                            lodash.isNumber(tag.childId)
+                        )
+                    ) {
+                        return false
+                    }
+                }
+            }
+        }
+
+        return true
+    }
+    const isValid = validateFlatTagGraph(tagGraph)
+    if (!isValid) {
+        throw new JsonError("Invalid tag graph provided", 400)
+    }
+    await db.updateTagGraph(trx, tagGraph)
+    res.send({ success: true })
+})
